New method for CMS logic and statements.
Before, there was no way to have something like `Definition 0` show up
for a definition within the document it appears, since importing the
Svelte component does not amount to connecting to the content layer.
Now, there is a Vite transformer which appends additional exports to
statement components, so that they may be rendered on the page in which
they are imported.